requirements.txt
Django>=2.2.0, <=2.2.4
djangorestframework>=3.10.0, <=3.10.2
psycopg2>=2.8.0, <=2.8.3
Pillow>=6.0.0, <=6.1.0

flake8>=3.7.0, <=3.7.8